Add additional testing for SqlAlchemyDataSet expectations
