+++ great_expectations/dataset/util.py
@@ -24,7 +24,10 @@ def parse_result_format(result_format):
             'result_obj_format': result_format,
             'partial_unexpected_count': 20
         }
-        
+    else:
+        if 'partial_unexpected_count' not in result_format:
+            result_format['partial_unexpected_count'] = 20
+
     return result_format
 
 class DotDict(dict):
